Fixes #359:Refactor MessageListItem.js in the components
