+++ src/components/ChatApp/HardwareComponent.js
